+++ src/main/java/Commands/Deadline.java
@@ -1,10 +1,13 @@
+package Commands;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task {
-    private String at;
+public class Deadline extends Task {
+
+    private String by;
     private LocalDate date;
     private LocalTime time;
     private String dateFormat;
@@ -22,17 +25,17 @@ public class Event extends Task {
             "HH:mm",
     };
 
-    public Event(String description, String at) {
-        super(description, "E");
-        this.at = at;
+    public Deadline(String description, String by) {
+        super(description, "D");
+        this.by = by;
     }
 
     public boolean isValidTime() {
         for (String format : timeFormats) {
             try {
-                if (this.at.split(" ").length != 1) {
+                if (this.by.split(" ").length != 1) {
                     DateTimeFormatter formatter = DateTimeFormatter.ofPattern(format);
-                    LocalTime.parse(this.at.split(" ")[1], formatter);
+                    LocalTime.parse(this.by.split(" ")[1], formatter);
                     this.timeFormat = format;
                     return true;
                 } else {
@@ -45,7 +48,7 @@ public class Event extends Task {
     }
 
     public boolean isValidDate() {
-        String d = this.at.split(" ")[0].replace("-", " ");
+        String d = this.by.split(" ")[0].replace("-", " ");
         for (String format : dateFormats) {
             try {
                 DateTimeFormatter formatter = DateTimeFormatter.ofPattern(format);
@@ -59,7 +62,7 @@ public class Event extends Task {
     }
     @Override
     public String getDate() {
-        String d = this.at.split(" ")[0].replace("-", " ");
+        String d = this.by.split(" ")[0].replace("-", " ");
         if (this.isValidDate()) {
             DateTimeFormatter formatter = DateTimeFormatter.ofPattern(this.dateFormat);
             this.date = LocalDate.parse(d, formatter);
@@ -72,15 +75,16 @@ public class Event extends Task {
     public String getTime() {
         if (this.isValidTime()) {
             DateTimeFormatter formatter = DateTimeFormatter.ofPattern(this.timeFormat);
-            this.time = LocalTime.parse(this.at.split(" ")[1], formatter);
+            this.time = LocalTime.parse(this.by.split(" ")[1], formatter);
             return this.time.format(DateTimeFormatter.ofPattern("HH:mm a"));
         } else {
             return "";
         }
     }
 
+
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: " + (this.getDate() + " " + this.getTime()).trim() + ")";
+        return "[D]" + super.toString() + " (by: " + (this.getDate() + " " + this.getTime()).trim() + ")";
     }
 }
